move the main model codes into a package namespace.
