+++ create_pretraining_data.py
@@ -20,7 +20,7 @@ from __future__ import print_function
 
 import collections
 import random
-import tokenization
+from bert import tokenization
 import tensorflow as tf
 
 flags = tf.flags
