+++ optimization_test.py
@@ -16,7 +16,7 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-import optimization
+from bert import optimization
 import tensorflow as tf
