+++ run_classifier_with_tfhub.py
@@ -19,9 +19,9 @@ from __future__ import division
 from __future__ import print_function
 
 import os
-import optimization
+from bert import optimization
+from bert import tokenization
 import run_classifier
-import tokenization
 import tensorflow as tf
 import tensorflow_hub as hub
